requirements.txt
Django==1.7.7
python-social-auth
django-bootstrap3
psycopg2
django-oscar
